refacto: moved the entire gui folder (this code is still a WIP though)
