+++ src/sports_calendar/interfaces/gui/app.py
@@ -1,7 +1,9 @@
+from .pages import root, selection_details
+
 from . import logger
-from .pages import root, selections, selection_details
+from .pages import selections
 from sports_calendar.core import Paths
-from sports_calendar.core.selection import SelectionService
+from sports_calendar.application.selection import SelectionService
 
 
 def create_app():
