+++ src/sports_calendar/interfaces/gui/cli.py
