+++ src/sports_calendar/interfaces/gui/components/filters/filter_modal.py
